+++ vignettes/examples/fcnresnet.R
@@ -5,8 +5,8 @@ library(torch)
 url1 <- "https://raw.githubusercontent.com/pytorch/vision/main/gallery/assets/dog1.jpg"
 url2 <- "https://raw.githubusercontent.com/pytorch/vision/main/gallery/assets/dog2.jpg"
 
-dog1 <- magick_loader(url1) |> transform_to_tensor()
-dog2 <- magick_loader(url2) |> transform_to_tensor()
+dog1 <- magick_loader(url1) %>% transform_to_tensor()
+dog2 <- magick_loader(url2) %>% transform_to_tensor()
 
 
 # Visualizing a grid of images -------------------------------------
@@ -23,11 +23,11 @@ tensor_image_browse(grid)
 norm_mean <- c(0.485, 0.456, 0.406)
 norm_std  <- c(0.229, 0.224, 0.225)
 
-dog1_prep <- dog1 |>
-  transform_resize(c(520,520)) |>
+dog1_prep <- dog1 %>%
+  transform_resize(c(520,520)) %>%
   transform_normalize(mean = norm_mean, std = norm_std)
-dog2_prep <- dog2 |>
-  transform_resize(c(520,520)) |>
+dog2_prep <- dog2 %>%
+  transform_resize(c(520,520)) %>%
   transform_normalize(mean = norm_mean, std = norm_std)
 
 # make batch (2,3,520,520)
@@ -54,13 +54,13 @@ mask$dtype
 
 
 segmented1 <- draw_segmentation_masks(
-  dog1 |> transform_resize(c(520,520)),
+  dog1 %>% transform_resize(c(520,520)),
   masks = mask[1,, ],
   alpha = 0.5
 )
 
 segmented2 <- draw_segmentation_masks(
-  dog2 |> transform_resize(c(520,520)),
+  dog2 %>% transform_resize(c(520,520)),
   masks = mask[2,, ],
   alpha = 0.5
 )